cmd: move waitFor logic into the pkg/wait
enables polling logic to be re-used when using installer as a package
